requirements.txt
absl-py==0.2.1
astor==0.6.2
backcall==0.1.0
bleach==1.5.0
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
gast==0.2.0
google-api-core==1.2.0
google-api-python-client==1.6.7
google-auth==1.4.1
google-cloud-bigquery==1.1.0
google-cloud-core==0.28.1
google-cloud-datastore==1.6.0
google-cloud-language==1.0.2
google-cloud-logging==1.6.0
google-cloud-pubsub==0.35.0
google-cloud-spanner==1.3.0
google-cloud-storage==1.10.0
google-cloud-translate==1.3.1
google-cloud-videointelligence==1.2.0
google-cloud-vision==0.31.1
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpc-google-iam-v1==0.11.4
grpcio==1.12.0
html5lib==0.9999999
httplib2==0.11.3
idna==2.6
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
Markdown==2.6.11
numpy==1.14.3
oauth2client==4.1.2
parso==0.2.1
pexpect==4.5.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycurl==7.43.0
Pygments==2.2.0
pygobject==3.22.0
python-apt==1.4.0b3
pytz==2018.4
requests==2.18.4
rsa==3.4.2
simplegeneric==0.8.1
six==1.11.0
speedtest-cli==2.0.2
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.22
virtualenv==16.0.0
wcwidth==0.1.7
Werkzeug==0.14.1
